+++ crucible-mir/src/Mir/TransCustom.hs
@@ -21,6 +21,7 @@
 {-# LANGUAGE FlexibleContexts #-}
 {-# LANGUAGE TupleSections #-}
 {-# LANGUAGE ConstraintKinds #-}
+{-# LANGUAGE LambdaCase #-}
 
 module Mir.TransCustom(customOps) where
 
@@ -164,6 +165,8 @@ customOpDefs = Map.fromList $ [
                          , ptr_swap
                          , ptr_null
                          , ptr_null_mut
+                         , drop_in_place_dyn
+
                          , intrinsics_copy
                          , intrinsics_copy_nonoverlapping
 
@@ -529,7 +532,10 @@ ptr_compare_usize = (["core", "crucible", "ptr", "compare_usize"],
                 let ptr = getSlicePtr slice
                 valAsPtr <- integerToMirRef val
                 MirExp C.BoolRepr <$> mirRef_eq ptr valAsPtr
-            -- TODO: `&dyn Tr` case (after defining MirDynRepr)
+            [MirExp DynRefRepr dynRef, MirExp UsizeRepr val] -> do
+                let ptr = S.getStruct dynRefDataIndex dynRef
+                valAsPtr <- integerToMirRef val
+                MirExp C.BoolRepr <$> mirRef_eq ptr valAsPtr
             _ -> mirFail $ "bad arguments for ptr::compare_usize: " ++ show ops
         _ -> Nothing)
 
@@ -623,6 +629,64 @@ null_ptr_impl what substs = case substs of
             return $ MirExp MirReferenceRepr ref
         _  -> mirFail $ "expected no arguments for " ++ what ++ ", received: " ++ show ops
     _ -> Nothing
+
+-- | Experimentally, we've observed that rustc seems not to generate proper drop
+-- glue for @&dyn Trait@ drops. We get around this by overriding
+-- @core::ptr::drop_in_place@ (when instantiated at @dyn@ types) to fetch the
+-- appropriate drop method from the trait object's vtable. @mir-json@ will have
+-- included this method.
+--
+-- For versions of @drop_in_place@ instantiated at other types, this override
+-- will not apply, and we will defer to the rustc-provided implementation.
+drop_in_place_dyn :: (ExplodedDefId, CustomRHS)
+drop_in_place_dyn =
+    ( ["core", "ptr", "drop_in_place"],
+      \case
+        Substs [TyDynamic traitName] ->
+            Just $ CustomOp $ \argTys [op@(MirExp selfTy selfExpr)] -> do
+                col <- use $ cs . collection
+                let argTys = Ctx.empty
+                let argExprs = Ctx.empty
+                let retTy = C.UnitRepr
+
+                -- We expect `mir-json` to have placed this trait object's drop
+                -- method at index 0, unless the trait object lacks a principal
+                -- trait (e.g. `dyn Send`). (This caveat for traits like `Send`
+                -- is a bug/limitation of `mir-json` - it _should_ emit a vtable
+                -- with a drop method for their trait objects, but that requires
+                -- some implementation effort we haven't yet spent, so it
+                -- currently does not.) Check here whether the trait has any
+                -- methods, to emit a more relevant error message than
+                -- `doVirtCall` would emit on its own.
+                let dropMethodIndex = 0
+                () <- case col ^. traits . at traitName of
+                    Nothing -> mirFail $ "undefined trait: "<>show traitName
+                    Just trait ->
+                        case trait ^. traitItems of
+                            [] -> mirFail $ "no drop method available in trait "<>show traitName
+                            (_:_) -> pure ()
+
+                callExpr <-
+                    doVirtCall
+                        col
+                        traitName
+                        dropMethodIndex
+                        selfTy
+                        selfExpr
+                        argTys
+                        argExprs
+                        retTy
+                pure (MirExp retTy callExpr)
+        Substs [_] ->
+            -- We weren't provided a `dyn`/`TyDynamic`, so we don't provide an
+            -- override, we just defer to the rustc-provided implementation.
+            Nothing
+        Substs ss ->
+            Just $ CustomOp $ \_ _ -> mirFail $ unwords
+                [ "expected one subst for `core::ptr::drop_in_place`, saw"
+                , show (length ss) <> ":"
+                , show ss ]
+    )
 
 intrinsics_copy :: (ExplodedDefId, CustomRHS)
 intrinsics_copy = ( ["core", "intrinsics", "copy"], \substs -> case substs of
@@ -1037,7 +1101,6 @@ mem_swap = (["core","mem", "swap"],
             return $ MirExp knownRepr $ R.App E.EmptyApp
         _ -> mirFail $ "bad arguments to mem_swap: " ++ show (opTys, ops)
     )
-
 
 -- This is like normal mem::transmute, but requires source and target types to
 -- have identical Crucible `TypeRepr`s.