+++ crux-mir/test/conc_eval/dyn/arc.rs
@@ -1,0 +1,33 @@
+//! Test introduction and dropping of `Arc<dyn Trait>`.
+
+use std::sync::Arc;
+
+trait Foo {}
+
+impl Foo for usize {}
+
+struct DropType(u32, u32);
+
+impl Foo for DropType {}
+
+static mut DROPPED: bool = false;
+
+impl Drop for DropType {
+    fn drop(&mut self) {
+        unsafe {
+            DROPPED = true;
+        }
+    }
+}
+
+#[cfg_attr(crux, crux::test)]
+fn crux_test() {
+    let x: Arc<dyn Foo> = Arc::new(3usize);
+    let y: Arc<dyn Foo> = Arc::new(DropType(3, 4));
+    drop(y);
+    assert!(unsafe { DROPPED });
+}
+
+fn main() {
+    println!("{:?}", crux_test());
+}